Separated model and manager modules in webiscite
